feat(pricelist): product branch pricing information https://unz.atlassian.net/wiki/spaces/TA/pages/1068663504
