+++ lib/index.js
@@ -1,6 +1,7 @@
 module.exports = {
   v0: require('./v0'),
   v1: require('./v1'),
+  v2: require('./v2'),
   v3: require('./v3'),
   common: require('./common')
 }
