+++ lib/v2/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  products: require('./products')
+}
