+++ lib/v2/products/index.js
@@ -1,0 +1,3 @@
+module.exports.list = {
+  response: require('./branch.response')
+}
